% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_directory.R
\name{check_directory}
\alias{check_directory}
\title{Check if there is a directory to use}
\usage{
check_directory(directory)
}
\arguments{
\item{directory}{Defaults to \code{NULL}. In this case the \code{ksh_data}
directory is used, if exists, or it is created.}
}
\description{
Check if there is a directory to use
}
\examples{
\dontrun{
check_directory ( directory = NULL )
}
}
