% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_megye_map.R
\name{create_megye_map}
\alias{create_megye_map}
\title{Create county map}
\usage{
create_megye_map(df, megyeid = "name", values_var = "values")
}
\arguments{
\item{df}{The data.frame with the values to map. The data must be in long-form
and contain only the values of a single point in time.}

\item{megyeid}{The name of the Hungarian counties, defaults to \code{megye}.}

\item{values_var}{The name of numeric values, defaults to \code{values}.}
}
\description{
Create county map
}
\examples{
\dontrun{
create_megye_map (df)
}
}
\keyword{choropleth}
\keyword{county,}
\keyword{hungary,}
\keyword{mapping,}
\keyword{megye,}
