% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stadat_file.R
\name{download_stadat_file}
\alias{download_stadat_file}
\title{Download a Stadat file}
\usage{
download_stadat_file(filename = "2_1_53i", directory = NULL)
}
\arguments{
\item{filename}{Currently defaults to \code{2_1_53i}.}

\item{directory}{Defaults to \code{NULL}. In this case the \code{ksh_data}
directory is used, if exists, or it is created.}
}
\description{
Download a Stadat file
}
\examples{
\dontrun{
download_stadat_file ( "2_1_53i" )
}
}
\keyword{hungary,}
\keyword{ksh,}
\keyword{opengov,}
\keyword{openstatistics}
