% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_real_income.R
\name{get_real_income}
\alias{get_real_income}
\title{Get Real Income (2.1.53i)
Get the Stadat 2.1.53 statistics}
\usage{
get_real_income(directory = NULL, class = NULL)
}
\arguments{
\item{directory}{Defaults to \code{NULL}.}

\item{class}{Defaults to \code{NULL}. If \code{ts} is specified, returns a
time series object.}
}
\description{
Get Real Income (2.1.53i)
Get the Stadat 2.1.53 statistics
}
\examples{
\dontrun{
ksh_real_income <- get_real_income ()
}
}
\keyword{hungary,}
\keyword{ksh,}
\keyword{opengov,}
\keyword{openstatistics}
