% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regional_active_population.R
\name{get_regional_active_population}
\alias{get_regional_active_population}
\title{Get economically activity population for Hungarian regions and counties}
\usage{
get_regional_active_population(directory = NULL, region_level = NULL)
}
\arguments{
\item{directory}{Defaults to \code{NULL}.}

\item{region_level}{Any of \code{megye}, \code{county},
\code{régió}, \code{region}, \code{nagyrégió}
or \code{NULL} for all data units.}
}
\description{
6.2.1.7. A gazdaságilag aktív népesség száma (2000–)
}
\examples{
\dontrun{
get_regional_gdp (region_level = "county")
}
}
\keyword{hungary,}
\keyword{ksh,}
\keyword{opengov,}
\keyword{openstatistics}
