% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regional_demography_age_group.R
\name{get_regional_demography_age_group}
\alias{get_regional_demography_age_group}
\title{Get resident population by age-group, for Hungarian regions and counties}
\source{
\url{https://www.ksh.hu/docs/eng/xstadat/xstadat_annual/i_wdsd004b.html}
}
\usage{
get_regional_demography_age_group(directory = NULL,
  region_level = NULL)
}
\arguments{
\item{directory}{Defaults to \code{NULL}.}

\item{region_level}{Any of \code{megye}, \code{county},
\code{régió}, \code{region}, \code{nagyrégió}, \code{large_region},
\code{national}, \code{ország}, or \code{NULL} for all data units.}
}
\description{
6.1.2. Resident population by age-group, 1 January (2001–)
}
\examples{
\dontrun{
get_regional_demography_age_group(region_level = "county")
}
}
\keyword{hungary,}
\keyword{ksh,}
\keyword{opengov,}
\keyword{openstatistics}
