% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regional_economic_activity_rate.R
\name{get_regional_economic_activity_rate}
\alias{get_regional_economic_activity_rate}
\title{Get economic activity rate for Hungarian regions and counties}
\source{
\url{https://www.ksh.hu/docs/eng/xstadat/xstadat_annual/i_qlf025b.html}
}
\usage{
get_regional_economic_activity_rate(directory = NULL,
  region_level = NULL)
}
\arguments{
\item{directory}{Defaults to \code{NULL}.}

\item{region_level}{Any of \code{megye}, \code{county},
\code{régió}, \code{region}, \code{nagyrégió}
or \code{NULL} for all data units.}
}
\description{
6.2.1.9. Activity rate (2000–)
}
\examples{
\dontrun{
get_regional_economic_activity_rate (region_level = "county")
}
}
\keyword{hungary,}
\keyword{ksh,}
\keyword{opengov,}
\keyword{openstatistics}
