% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regional_gdp.R
\name{get_regional_gdp}
\alias{get_regional_gdp}
\title{Get regional GDP data for Hungary}
\source{
\url{https://www.ksh.hu/docs/eng/xstadat/xstadat_annual/i_qpt012b.html}
}
\usage{
get_regional_gdp(directory = NULL, region_level = NULL)
}
\arguments{
\item{directory}{Defaults to \code{NULL}.}

\item{region_level}{Either \code{megye} or default \code{NULL} for all
data units.}
}
\description{
6.3.1.1. Gross domestic product (GDP) (2000–)
}
\examples{
\dontrun{
ksh_real_income <- get_real_income ()
}
}
\keyword{hungary,}
\keyword{ksh,}
\keyword{opengov,}
\keyword{openstatistics}
