% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regional_net_earnings.R
\name{get_regional_net_earnings}
\alias{get_regional_net_earnings}
\title{Average monthly net earnings of full-time employees by the geographic area of Hungary}
\source{
\url{https://www.ksh.hu/docs/eng/xstadat/xstadat_annual/i_qli050b.html}
}
\usage{
get_regional_net_earnings(directory = NULL, region_level = NULL)
}
\arguments{
\item{directory}{Defaults to \code{NULL}.}

\item{region_level}{Any of \code{megye}, \code{county},
\code{régió}, \code{region}, \code{nagyrégió}
or \code{NULL} for all data units.}
}
\description{
6.2.1.16. Average monthly net earnings of full-time employees by the geographic area of the local unit (2000–)
}
\examples{
\dontrun{
get_regional_net_earnings (region_level = "region")
}
}
\keyword{hungary,}
\keyword{ksh,}
\keyword{opengov,}
\keyword{openstatistics}
