% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regional_theatre_cinema_admissions.R
\name{get_regional_theatre_cinema_admissions}
\alias{get_regional_theatre_cinema_admissions}
\title{Get theatre and cinema admissions for Hungarian regions and counties}
\source{
\url{https://www.ksh.hu/docs/eng/xstadat/xstadat_annual/i_zkk004b.html}
}
\usage{
get_regional_theatre_cinema_admissions(directory = NULL,
  region_level = NULL)
}
\arguments{
\item{directory}{Defaults to \code{NULL}.}

\item{region_level}{Any of \code{megye}, \code{county},
\code{régió}, \code{region}, \code{nagyrégió}, \code{large_region},
or \code{NULL} for all data units.}
}
\description{
6.2.6.2. Number of theatre and cinema admissions (2000–)
}
\examples{
\dontrun{
get_regional_theatre_cinema_admissions (region_level = "county")
}
}
\keyword{hungary,}
\keyword{ksh,}
\keyword{opengov,}
\keyword{openstatistics}
