% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regional_dwellings.R
\name{get_regional_dwellings}
\alias{get_regional_dwellings}
\title{Get number of new dwellings and  for Hungarian regions and counties}
\source{
\url{https://www.ksh.hu/docs/eng/xstadat/xstadat_annual/i_wde003b.html}
}
\usage{
get_regional_dwellings(directory = NULL, region_level = NULL)
}
\arguments{
\item{directory}{Defaults to \code{NULL}.}

\item{region_level}{Any of \code{megye}, \code{county},
\code{régió}, \code{region}, \code{nagyrégió}
or \code{NULL} for all data units.}
}
\description{
6.2.2.1. Dwelling stock and inhabitants per hundred dwellings, 1 January (2001–)
}
\examples{
\dontrun{
get_regional_dwellings (region_level = "county")
}
}
\keyword{hungary,}
\keyword{ksh,}
\keyword{opengov,}
\keyword{openstatistics}
