% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regional_gdp.R
\name{get_regional_gdp}
\alias{get_regional_gdp}
\title{Get regional GDP data for Hungary}
\usage{
get_regional_gdp(directory = NULL, region_level = NULL, class = NULL)
}
\arguments{
\item{directory}{Defaults to \code{NULL}.}

\item{class}{Defaults to \code{NULL}. If \code{ts} is specified, returns a
time series object.}
}
\description{
Get the Stadat 6.3.1.1i statistics
}
\examples{
\dontrun{
ksh_real_income <- get_real_income ()
}
}
\keyword{hungary,}
\keyword{ksh,}
\keyword{opengov,}
\keyword{openstatistics}
